Move basecommand out of commands folder to not pollute the help menu
