+++ lib/commands/shinject.py
@@ -12,7 +12,7 @@ __description__ = "Execute shellcode inside a process"
 __author__ = "@_batsec_"
 
 # identify the task as shellcode execute
-from lib.commands.basecommand import BaseCommand
+from lib.basecommand import BaseCommand
 
 DLLINJECT_EXEC_ID = 0x5000
